[feat] mock API 구현
