--- backend/src/main/java/com/codesquad/sidedish/SideDish/controller/CategoryController.java
@@ -1,26 +1,0 @@
-package com.codesquad.sidedish.SideDish.controller;
-
-import com.codesquad.sidedish.SideDish.dto.CategoryDto;
-import com.codesquad.sidedish.SideDish.service.CategoryService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/category")
-public class CategoryController {
-    private final CategoryService categoryService;
-
-    CategoryController(CategoryService categoryService) {
-        this.categoryService = categoryService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<CategoryDto>> getCategories() {
-        List<CategoryDto> categories = categoryService.getList();
-        return ResponseEntity.ok().body(categories);
-    }
-}
