+++ backend/src/main/java/com/codesquad/sidedish/SideDish/service/CategoryService.java
@@ -1,23 +1,18 @@
 package com.codesquad.sidedish.SideDish.service;
 
 import com.codesquad.sidedish.SideDish.dto.CategoryDto;
-import com.codesquad.sidedish.SideDish.dto.DishDto;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 @Service
 public class CategoryService {
     public List<CategoryDto> getList() {
-        List<DishDto> mainDishes = new ArrayList<>();
-        List<DishDto> soupDishes = new ArrayList<>();
-        List<DishDto> sideDishes = new ArrayList<>();
         return Arrays.asList(
-                new CategoryDto(1, "main", mainDishes),
-                new CategoryDto(2, "soup", soupDishes),
-                new CategoryDto(3, "side", sideDishes)
+                new CategoryDto(1, "메인", "/main"),
+                new CategoryDto(2, "국", "/soup"),
+                new CategoryDto(3, "반찬", "side")
         );
     }
 }
